requirements.txt
BAC0~=21.12.3
bitstring==4.0.1
boto3==1.26.86
coverage==7.2.2
cryptography==39.0.2
ecdsa==0.18.0
esp_docs==1.4.1
matplotlib==3.7.1
mysql_connector_repackaged==0.3.1
pkcs11==1.0.4
pyelftools==0.29
pyserial==3.5
pytest==7.2.2
python-dotenv==1.0.0
pytz==2023.3
reedsolo==1.6.0
setuptools==67.6.1

mysql~=0.0.3
mysql-connector-python~=8.0.32
esptool~=4.5.1